backend/requirement.txt
Package                  Version
------------------------ ----------
aiohappyeyeballs         2.6.1
aiohttp                  3.13.2
aiosignal                1.4.0
annotated-doc            0.0.3
annotated-types          0.7.0
anyio                    4.11.0
async-timeout            4.0.3
attrs                    25.4.0
certifi                  2025.10.5
charset-normalizer       3.4.4
click                    8.3.0
dataclasses-json         0.6.7
defusedxml               0.7.1
exceptiongroup           1.3.0
faiss-cpu                1.12.0
fastapi                  0.120.4
filelock                 3.20.0
frozenlist               1.8.0
fsspec                   2025.10.0
greenlet                 3.2.4
h11                      0.16.0
hf-xet                   1.2.0
httpcore                 1.0.9
httpx                    0.28.1
httpx-sse                0.4.3
huggingface-hub          0.36.0
idna                     3.11
Jinja2                   3.1.6
joblib                   1.5.2
jsonpatch                1.33
jsonpointer              3.0.0
langchain                1.0.3
langchain-classic        1.0.0
langchain-community      0.4.1
langchain-core           1.0.2
langchain-huggingface    1.0.0
langchain-text-splitters 1.0.0
langgraph                1.0.2
langgraph-checkpoint     3.0.0
langgraph-prebuilt       1.0.2
langgraph-sdk            0.2.9
langsmith                0.4.39
MarkupSafe               2.1.5
marshmallow              3.26.1
mpmath                   1.3.0
multidict                6.7.0
mypy_extensions          1.1.0
networkx                 3.3
numpy                    2.2.6
orjson                   3.11.4
ormsgpack                1.11.0
packaging                25.0
pillow                   12.0.0
pip                      22.0.2
propcache                0.4.1
pydantic                 2.12.3
pydantic_core            2.41.4
pydantic-settings        2.11.0
python-dotenv            1.2.1
PyYAML                   6.0.3
regex                    2025.10.23
requests                 2.32.5
requests-toolbelt        1.0.0
safetensors              0.6.2
scikit-learn             1.7.2
scipy                    1.15.3
sentence-transformers    5.1.2
setuptools               59.6.0
sniffio                  1.3.1
SQLAlchemy               2.0.44
starlette                0.49.3
sympy                    1.14.0
tenacity                 9.1.2
threadpoolctl            3.6.0
tokenizers               0.22.1
torch                    2.9.0+cpu
tqdm                     4.67.1
transformers             4.57.1
typing_extensions        4.15.0
typing-inspect           0.9.0
typing-inspection        0.4.2
urllib3                  2.5.0
uvicorn                  0.38.0
xxhash                   3.6.0
yarl                     1.22.0
youtube-transcript-api   0.6.2
yt-dlp                   2025.10.22
zstandard                0.25.0
